+++ jovo-integrations/jovo-platform-dialogflow/src/DialogflowAgent.ts
@@ -1,16 +1,16 @@
 import {BaseApp, Jovo, Host, SpeechBuilder} from "jovo-core";
 import {DialogflowUser} from "./DialogflowUser";
 import _get = require('lodash.get');
+import _set = require('lodash.set');
+
+export type SupportedIntegration = 'FacebookMessenger' | 'Slack';
 
 export class DialogflowAgent extends Jovo {
+    $dialogflowAgent: DialogflowAgent;
 
     constructor(app: BaseApp, host: Host) {
         super(app, host);
-        // this.$request = DialogflowRequest.fromJSON(hostwrapper.getRequestObject()) as DialogflowRequest;
-        // this.$request.requestObj = hostwrapper.getRequestObject();
-        // this.$response = new DialogflowResponse();
-        // this.$sessionAttributes = this.$request.getSessionAttributes();
-        // this.$inputs = this.$request.getInputs();
+        this.$dialogflowAgent = this;
         this.$user = new DialogflowUser(this);
     }
 
@@ -138,4 +138,7 @@ export class DialogflowAgent extends Jovo {
         return undefined;
     }
 
+    setCustomPayload(platform: string, payload: object) {
+        _set(this.$output, 'Dialogflow.Payload.' + platform, payload);
+    }
 }