+++ jovo-integrations/jovo-platform-googleassistant/src/modules/Transaction.ts
@@ -1,10 +1,12 @@
-import {EnumRequestType, Plugin, SpeechBuilder} from "jovo-core";
+import {EnumRequestType, Plugin, SpeechBuilder, ErrorCode, JovoError} from "jovo-core";
 import _set = require('lodash.set');
 import _get = require('lodash.get');
 
 import {GoogleAssistant} from "../GoogleAssistant";
 import {GoogleAction} from "../core/GoogleAction";
 import {GoogleActionResponse} from "../core/GoogleActionResponse";
+import {GoogleActionAPI} from "../services/GoogleActionAPI";
+import {GoogleActionAPIResponse} from "../services/GoogleActionAPIResponse";
 
 export interface PaymentOptions {
     googleProvidedOptions: GoogleProvidedOptions;
@@ -36,8 +38,8 @@ export interface Requirements {
 export type RequirementsCheckResult = 'USER_ACTION_REQUIRED' | 'OK' | 'ASSISTANT_SURFACE_NOT_SUPPORTED' | 'REGION_NOT_SUPPORTED';
 export type DeliveryAddressDecision = 'ACCEPTED' | 'REJECTED';
 export type TransactionDecision = 'ORDER_ACCEPTED' | 'REJECTED' | 'ORDER_REJECTED' | 'DELIVERY_ADDRESS_UPDATED' | 'CART_CHANGE_REQUESTED';
-
-
+export type SkuType = 'SKU_TYPE_IN_APP' | 'SKU_TYPE_SUBSCRIPTION';
+export type PurchaseStatus = 'PURCHASE_STATUS_OK' | 'PURCHASE_STATUS_ITEM_CHANGE_REQUESTED' | 'PURCHASE_STATUS_USER_CANCELLED' | 'PURCHASE_STATUS_ERROR' | 'PURCHASE_STATUS_UNSPECIFIED';
 
 export interface DeliveryAddressLocation {
     zipCode?: string;
@@ -72,11 +74,12 @@ export interface OrderUpdate {
 
 export class Transaction {
     googleAction: GoogleAction;
+    googleAssistant: GoogleAssistant;
 
-    constructor(googleAction: GoogleAction) {
+    constructor(googleAction: GoogleAction, googleAssistant: GoogleAssistant) {
         this.googleAction = googleAction;
+        this.googleAssistant = googleAssistant;
     }
-
 
 
     // REQUIREMENTS CHECK
@@ -293,14 +296,109 @@ export class Transaction {
         return this.getTransactionDecisionResult() === "CART_CHANGE_REQUESTED";
     }
 
+
+    getSubscriptions(skus: string[]): Promise<any[]> { // tslint:disable-line
+        return this.getSkus(skus, "SKU_TYPE_SUBSCRIPTION");
+    }
+
+    getConsumables(skus: string[]): Promise<any[]> { // tslint:disable-line
+        return this.getSkus(skus, "SKU_TYPE_IN_APP");
+    }
+
+    completePurchase(skuId: string) {
+        this.googleAction.$output.GoogleAssistant = {
+            CompletePurchase: {
+                skuId,
+            }
+        };
+    }
+    getPurchaseStatus(): PurchaseStatus | undefined {
+        for (const argument of _get(this.googleAction.$originalRequest || this.googleAction.$request, 'inputs[0]["arguments"]', [])) {
+            if (argument.name === 'COMPLETE_PURCHASE_VALUE') {
+                return argument.extension.purchaseStatus;
+            }
+        }
+    }
+
+    private getSkus(skus: string[], type: SkuType): Promise<any[]> { // tslint:disable-line
+        const conversationId = _get(this.googleAction.$request, 'originalDetectIntentRequest.payload.conversation.conversationId');
+
+        if (!this.googleAssistant.config.transactions || !this.googleAssistant.config.transactions.androidPackageName) {
+            throw new JovoError(
+                'getSkus needs the Android App package name',
+                ErrorCode.ERR,
+                'jovo-platform-googleassistant'
+            );
+        }
+
+        const promise = this.getGoogleApiAccessToken()
+            .then((accessToken: string) => {
+                return GoogleActionAPI.apiCall({
+                    endpoint: 'https://actions.googleapis.com',
+                    path: `/v3/packages/${this.googleAssistant.config.transactions!.androidPackageName}/skus:batchGet`,
+                    method: 'POST',
+                    permissionToken: accessToken,
+                    json: {
+                        conversationId,
+                        skuType: type,
+                        ids: skus
+                    }
+                }) as Promise<GoogleActionAPIResponse>;
+            });
+
+        return promise.then((response: GoogleActionAPIResponse) => response.data);
+    }
+
+    private getGoogleApiAccessToken(): Promise<string> {
+        if (!this.googleAssistant.config.transactions || !this.googleAssistant.config.transactions.keyFile) {
+            throw new JovoError(
+                'Please add a valid keyFile object to the GoogleAssistant transaction config',
+                ErrorCode.ERR,
+                'jovo-platform-googleassistant'
+            );
+        }
+        /**
+         * DigitalGoods.ts needs the googleapis package to function.
+         * To reduce overall package size, googleapis wasn't added as a dependency.
+         * googleapis has to be manually installed
+         */
+        try {
+            const googleapis = require('googleapis').google;
+
+            return googleapis.auth.getClient({
+                keyFile: this.googleAssistant.config.transactions!.keyFile,
+                scopes: [ 'https://www.googleapis.com/auth/actions.purchases.digital' ]
+            })
+                .then((client: any) => client.authorize()) // tslint:disable-line
+                .then((authorization: any) => authorization.access_token as string); // tslint:disable-line
+
+        } catch (e) {
+            console.log(e);
+            console.log(e.stack);
+            if (e.message === 'Cannot find module \'googleapis\'') {
+                return Promise.reject(new JovoError(
+                    e.message,
+                    ErrorCode.ERR,
+                    'jovo-platform-googleassistant',
+                    undefined,
+                    'Please run `npm install googleapis`'
+                ));
+            }
+            return Promise.reject(new Error('Could not retrieve Google API access token'));
+        }
+
+    }
+
 }
 
 
 export class TransactionsPlugin implements Plugin {
+    googleAssistant?: GoogleAssistant;
 
     install(googleAssistant: GoogleAssistant) {
         googleAssistant.middleware('$type')!.use(this.type.bind(this));
         googleAssistant.middleware('$output')!.use(this.output.bind(this));
+        this.googleAssistant = googleAssistant;
 
         GoogleAction.prototype.$transaction = undefined;
     }
@@ -321,10 +419,14 @@ export class TransactionsPlugin implements Plugin {
         if (_get(googleAction.$originalRequest || googleAction.$request, 'inputs[0].intent') === 'actions.intent.TRANSACTION_DECISION') {
             _set(googleAction.$type, 'type', 'ON_TRANSACTION');
             _set(googleAction.$type, 'subType', 'TRANSACTION_DECISION');
-
         }
 
-        googleAction.$transaction = new Transaction(googleAction);
+
+        if (_get(googleAction.$originalRequest || googleAction.$request, 'inputs[0].intent') === 'actions.intent.TRANSACTION_DECISION') {
+            _set(googleAction.$type, 'type', 'ON_TRANSACTION');
+            _set(googleAction.$type, 'subType', 'COMPLETE_PURCHASE');
+        }
+        googleAction.$transaction = new Transaction(googleAction, this.googleAssistant!);
     }
 
     output(googleAction: GoogleAction) {
@@ -393,9 +495,28 @@ export class TransactionsPlugin implements Plugin {
                 }
             });
             _set(googleAction.$response, 'richResponse.items', richResponseItems);
-
         }
 
+        const completePurchase = _get(output, "GoogleAssistant.CompletePurchase");
+
+        if (completePurchase) {
+            _set(googleAction.$response, "expectUserResponse", true);
+            _set(googleAction.$response, "systemIntent", {
+                intent: "actions.intent.COMPLETE_PURCHASE",
+                inputValueData: {
+                    "@type": "type.googleapis.com/google.actions.transactions.v3.CompletePurchaseValueSpec",
+                    skuId: completePurchase.skuId
+                },
+            });
+            _set(googleAction.$response, 'inputPrompt', {
+                initialPrompts: [
+                    {
+                        textToSpeech: 'PLACEHOLDER',
+                    },
+                ],
+                noInputPrompts: [],
+            });
+        }
     }
     uninstall(googleAssistant: GoogleAssistant) {